backend/requirements.txt
fastapi
uvicorn
langchain
langgraph
langchain-openai
pydantic
python-dotenv
requests
